test: more reliable decoder controller test
